+++ tests/test_core.py
@@ -38,6 +38,140 @@ def test_decorators():
     assert should_be_a_foreach_recipe.transient is False
 
 
+def test_parse_docstring_from_func() -> None:
+    """
+    Test the '_parse_docstring_from_func' function is able to parse function descriptions from various docstring formats
+    """
+    # Test that an empty docstring doesn't blow up
+
+    assert alk.decorators._parse_docstring_from_func(lambda x: x ** 2) == ""
+
+    def square(x: float) -> float:
+        return x ** 2
+
+    assert alk.decorators._parse_docstring_from_func(square) == ""
+
+    # Test that a lambda with a preceding comment works
+
+    # Square the provided value
+    square_lambda = lambda x: x ** 2  # NOQA: Testing only
+    assert alk.decorators._parse_docstring_from_func(square_lambda) == "Square the provided value"
+
+    # Test that a single-line docstring works
+    def square_single_line(x: float) -> float:
+        """Square the provided value"""
+        return x ** 2
+
+    assert alk.decorators._parse_docstring_from_func(square_single_line) == "Square the provided value"
+
+    # Test that a reST docstring works
+    def square_rest(x: float) -> float:
+        """
+        Square the provided value
+
+        :param x: The value to square
+        :return: The squared value
+        """
+        return x ** 2
+
+    assert alk.decorators._parse_docstring_from_func(square_rest) == "Square the provided value"
+
+    # Test that a multiline reST docstring works
+    def square_rest_multiline(x: float) -> float:
+        """
+        Square the provided value
+        and return the result
+
+        :param x: The value to square
+        :return: The squared value
+        """
+        return x ** 2
+
+    assert alk.decorators._parse_docstring_from_func(
+        square_rest_multiline) == "Square the provided value and return the result"
+
+    # Test that a reST docstring works, even without a blank newline between description and params
+    def square_rest_no_newline(x: float) -> float:
+        """
+        Square the provided value
+        :param x: The value to square
+        :return: The squared value
+        """
+        return x ** 2
+
+    assert alk.decorators._parse_docstring_from_func(square_rest_no_newline) == "Square the provided value"
+
+    # Test that a google docstring works
+    def square_google(x: float) -> float:
+        """
+        Square the provided value
+
+        Args:
+            x: The value to square.
+
+        Returns:
+            The squared value
+        """
+        return x ** 2
+
+    assert alk.decorators._parse_docstring_from_func(square_google) == "Square the provided value"
+
+    # Test that a numpydoc docstring works
+    def square_numpydoc(x: float) -> float:
+        """
+        Square the provided value
+
+        Parameters
+        ----------
+        x : float
+            The value to square
+
+        Returns
+        -------
+        float
+            The squared value
+        """
+        return x ** 2
+
+    assert alk.decorators._parse_docstring_from_func(square_numpydoc) == "Square the provided value"
+
+
+def test_docstrings() -> None:
+    """
+    Test that the decorators parse docstrings correctly, but prefers user-provided values
+    """
+
+    @alk.recipe(doc="This is overridden")
+    def func_with_overridden_doc() -> None:
+        pass
+
+    assert func_with_overridden_doc.doc == "This is overridden"
+
+    @alk.recipe()
+    def func() -> None:
+        """
+        A proper docstring for this function
+        """
+        pass
+
+    assert func.doc == "A proper docstring for this function"
+
+    @alk.foreach(func, doc="This is overridden")
+    def foreach_with_overridden_doc() -> None:
+        pass
+
+    assert foreach_with_overridden_doc.doc == "This is overridden"
+
+    @alk.foreach(func)
+    def foreach() -> None:
+        """
+        A proper docstring for this function
+        """
+        pass
+
+    assert foreach.doc == "A proper docstring for this function"
+
+
 def test_brew():
     AlkymiConfig.get().cache = False