+++ tests/test_utils.py
@@ -34,7 +34,8 @@ def test_call_error():
     assert error_str in stream.read()
 
 
-def test_call_updates():
+@pytest.mark.parametrize("stream_to_forward", ("stdout", "stderr"))
+def test_call_updates(stream_to_forward: str):
     """
     Test that the utils.call() function can correctly return stdout line-by-line while a command is executing (to enable
     live output).
@@ -64,16 +65,20 @@ def test_call_updates():
     stream_monitor.start()
 
     # Script to execute - simply prints a number of numbers, each on a separate line followed by a sleep
-    program = """
+    program = f"""
 import time
+import sys
 
-for i in range({}):
-    print(i)
+for i in range({num_outputs}):
+    print(i, file=sys.{stream_to_forward})
     time.sleep(0.1)
-    """.format(num_outputs)
+    """
 
     # Pass '-u' to avoid buffering inside the subprocess itself
-    alkymi.utils.call(["python3", "-u", "-c", program], echo_output_to_stream=stream)
+    if stream_to_forward == "stdout":
+        alkymi.utils.call(["python3", "-u", "-c", program], echo_output_to_stream=stream)
+    elif stream_to_forward == "stderr":
+        alkymi.utils.call(["python3", "-u", "-c", program], echo_error_to_stream=stream)
 
     # Stop reader thread
     process_stream = False
@@ -103,7 +108,7 @@ def test_run_on_thread():
     assert _call() == current_thread_idx
 
     # Running "call" on a different thread should result in another ID
-    assert alkymi.utils.run_on_thread(_call) != current_thread_idx
+    assert alkymi.utils.run_on_thread(_call)() != current_thread_idx
 
 
 def test_run_on_thread_none():
@@ -114,7 +119,7 @@ def test_run_on_thread_none():
     def _call() -> None:
         pass
 
-    alkymi.utils.run_on_thread(_call)
+    alkymi.utils.run_on_thread(_call)()
 
 
 def test_run_on_thread_exception():
@@ -127,12 +132,12 @@ def test_run_on_thread_exception():
 
     # Running "call" on a different thread should result in another ID
     with pytest.raises(TypeError):
-        alkymi.utils.run_on_thread(_fail)
+        alkymi.utils.run_on_thread(_fail)()
 
 
 def test_check_current_thread_has_running_event_loop():
     # Running in a new thread should cause the check to return false due to no running event loop in the new thread
-    assert not alkymi.utils.run_on_thread(lambda: alkymi.utils.check_current_thread_has_running_event_loop())
+    assert not alkymi.utils.run_on_thread(lambda: alkymi.utils.check_current_thread_has_running_event_loop())()
 
     def _call_async() -> bool:
         loop = asyncio.new_event_loop()
@@ -144,4 +149,4 @@ def test_check_current_thread_has_running_event_loop():
         return loop.run_until_complete(_wrap())
 
     # Running in a new thread with a new event loop should cause the function to return true
-    assert alkymi.utils.run_on_thread(_call_async)
+    assert alkymi.utils.run_on_thread(_call_async)()